# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SwiftEntry API",
        "description": "Backend REST de SwiftEntry \u2014 plataforma de venta de boletos para eventos.\n\n**Autenticaci\u00f3n:** todas las rutas protegidas requieren un JWT en el header `Authorization: Bearer <token>`. Obt\u00e9n el token en `POST /swift_entry/auth/login` y p\u00e9galo en el bot\u00f3n **Authorize** de arriba.\n\n**Flujo principal:** Evento \u2192 Localidad \u2192 Asiento \u2192 Reserva \u2192 Pago \u2192 Boleto\n",
        "contact": {
            "name": "SwiftEntry Team",
            "email": "gerardo.port1107.psql@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/swift_entry/auth/refresh": {
            "post": {
                "tags": [
                    "Autenticaci\u00f3n"
                ],
                "summary": "Renovar token",
                "description": "Renueva el access token usando el refresh token. El refresh token se rota en cada llamada",
                "operationId": "refresh",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/swift_entry/auth/logout": {
            "post": {
                "tags": [
                    "Autenticaci\u00f3n"
                ],
                "summary": "Cerrar sesi\u00f3n",
                "description": "Invalida el refresh token en base de datos",
                "operationId": "logout",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/swift_entry/auth/login": {
            "post": {
                "tags": [
                    "Autenticaci\u00f3n"
                ],
                "summary": "Iniciar sesi\u00f3n",
                "description": "Retorna un access token (15 min) y un refresh token (7 d\u00edas)",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GeneralResponse": {
                "type": "object",
                "properties": {
                    "uri": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {}
                }
            },
...
            "RefreshRequestDTO": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "refreshToken"
                ]
            },
            "AuthRequestDTO": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}